move provider to vscodeInlineCompletionItemProvider.ts
