+++ vscode/src/completions/providers/unstable-azure-openai.ts
@@ -1,7 +1,7 @@
-import { Completion } from '..'
 import { logger } from '../../log'
 import { ReferenceSnippet } from '../context'
 import { getHeadAndTail } from '../text-processing'
+import { Completion } from '../vscodeInlineCompletionItemProvider'
 
 import { Provider, ProviderConfig, ProviderOptions } from './provider'
