+++ vscode/src/completions/providers/unstable-fireworks.ts
@@ -1,10 +1,10 @@
 import fetch from 'isomorphic-fetch'
 
-import { Completion } from '..'
 import { logger } from '../../log'
 import { ReferenceSnippet } from '../context'
 import { getLanguageConfig } from '../language'
 import { isAbortError } from '../utils'
+import { Completion } from '../vscodeInlineCompletionItemProvider'
 
 import { Provider, ProviderConfig, ProviderOptions } from './provider'
