+++ vscode/src/completions/request-manager.ts
@@ -1,8 +1,8 @@
-import { Completion } from '.'
 import { CompletionsCache } from './cache'
 import { ReferenceSnippet } from './context'
 import { logCompletionEvent } from './logger'
 import { CompletionProviderTracer, Provider } from './providers/provider'
+import { Completion } from './vscodeInlineCompletionItemProvider'
 
 interface Request {
     prefix: string
