+++ vscode/src/completions/shared-post-process.ts
@@ -1,6 +1,6 @@
-import { Completion } from '.'
 import { truncateMultilineCompletion } from './multiline'
 import { collapseDuplicativeWhitespace, trimUntilSuffix } from './text-processing'
+import { Completion } from './vscodeInlineCompletionItemProvider'
 
 /**
  * This function implements post-processing logic that is applied regardless of
